Support BigInteger, generate 'if' for non-int switch
