+++ test/resources/pythag/pythag.idr
@@ -2,7 +2,7 @@ module Main
 
 pythag : Int -> List (Int, Int, Int)
 pythag max = [(x, y, z) | z <- [1..max], y <- [1..z], x <- [1..y],
-                          x * x + y *y == z * z]
+                          x * x + y * y == z * z]
 
 main : IO ()
 main = printLn (pythag 50)
